web/requirements.txt
Django==1.8.1
django-simple-captcha
gunicorn==19.3.0
psycopg2==2.6
